requirements.txt
nfstream==3.2.2
pandas==1.0.3
scapy==2.4.3
zat==0.3.8
# for 'local' stuff :
# click
google-cloud-bigquery==1.24.0